wip: event page

+++ src/main/java/br/com/newtonpaiva/fastpass/dto/DashboardResponseDTO.java
@@ -4,13 +4,11 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @Builder
 public class DashboardResponseDTO {
-    private BigDecimal balanceAvailable;
+    private Double balanceAvailable;
     private Long pastEvents;
     private Long futureEvents;
     private Long rechargesPerformed;
